requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
aiomysql==0.2.0

# Auth and security
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.17

# Validation and config
pydantic[email]==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.2.0
